# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-Commerce Analytics & Recommendation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations/personalized/{userId}:
    get:
      operationId: getPersonalizedRecommendations
      tags: [recommendations]
      summary: Personalized product recommendations for user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRecommendation"
  /recommendations/similar/{productId}:
    get:
      operationId: getSimilarProducts
      tags: [recommendations]
      summary: Similar products
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRecommendation"
  /recommendations/trending:
    get:
      operationId: getTrendingProducts
      tags: [recommendations]
      summary: Trending products
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Trending products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRecommendation"
  /recommendations/bought-together/{productId}:
    get:
      operationId: getFrequentlyBoughtTogether
      tags: [recommendations]
      summary: Frequently bought together products
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Frequently bought together
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRecommendation"

  # FORECASTING
...
components:
  schemas:
...
    ProductRecommendation:
      type: object
      required: [id, name, category, price, rating, score, reason]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        rating:
          type: number
        score:
          type: number
        reason:
          type: string
        imageUrl:
          type: ["string", "null"]
...