# lib/api-zod/src/generated/types/productRecommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Analytics & Recommendation System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductRecommendation {
  id: number;
  name: string;
  category: string;
  price: number;
  rating: number;
  score: number;
  reason: string;
  /** @nullable */
  imageUrl?: string | null;
}